requirements.txt
selenium
tqdm
python-dotenv